+++ app/Http/Controllers/Crypto/OfferController.php
@@ -12,29 +12,210 @@ class OfferController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth', ['except' => ['index', 'show']]);
+        $this->middleware('auth:sanctum', ['except' => ['index', 'show']]);
     }
 
-    private function TradeOffer(int $crypto, int $user, int $price, int $amount, bool $selling, ?int $offer = null) {
+    private function TradeOffer(int $crypto, int $user, int $price, int $amount, bool $selling, ?int $offer_id = null) {
         $sort1 = '';
         $sort2 = '';
-        switch ($selling) {
-            case true:
-                $sort1 = '>=';
-                $sort2 = 'desc';
-                break;
-            
-            case false:
-                $sort1 = '<=';
-                $sort2 = 'asc';
-                break;
+        if ($selling) {
+            $sort1 = '>=';
+            $sort2 = 'desc';
+            if(isset($offer_id)) {
+
+            } else {
+
+            }
+        } else {
+            $sort1 = '<=';
+            $sort2 = 'asc';
+            if(isset($offer_id)) {
+
+            } else {
+                
+            }
         }
-        $offers = DB::select("SELECT `id`,`price`,`amount` from `offers` where `crypto_id` = ? and `selling` = ? and price ? order by `price` ?", [
-            $crypto,
-            !$selling,
-            $sort1.' '.$price,
-            $sort2,
-        ]);
+        $traded = 0;
+        $back = 0;
+        $count = 1;
+        while ($amount > 0 && $count > 0) {
+            if ($selling) {
+                $offer = DB::select("SELECT `offers`.`id`,`offers`.`price`,`offers`.`amount`,`offers`.`user_id`,`cryptoBalance`.`balance`,`users`.`balance` as `user_balance` from `offers`
+                left join `users`
+                on `users`.`id` = `offers`.`user_id`
+                left join (
+                    SELECT user_id, MAX(`c_t_id`) as `CT`
+                    FROM cryptoBalance
+                    WHERE crypto_id = ?
+                    GROUP BY user_id
+                ) r
+                on r.user_id = offers.user_id
+                left join `cryptoBalance`
+                on `cryptoBalance`.`user_id` = `offers`.`user_id` and `cryptoBalance`.`crypto_id` = `offers`.`crypto_id` and `cryptoBalance`.`c_t_id` = r.`CT`
+                where `offers`.`crypto_id` = ? and `offers`.`selling` = ? and `offers`.`user_id` != ? and `offers`.`price` $sort1 ? order by `offers`.`price` $sort2 limit 1", [
+                $crypto,
+                $crypto,
+                !$selling,
+                $user,
+                $price,
+                ]);
+                $count = count($offer);
+                if ($count > 0) {
+                    $offer = $offer[0];
+                    $changeofferamount = $offer['amount'];
+                    if ($changeofferamount > $amount) {
+                        $changeofferamount = $amount;
+                    }
+                    $id = DB::table('crypto_transactions')->insertGetId(
+                        ['in_out' => true, 'amount' => $changeofferamount, 'state' => 'success',
+                        'user_id' => $offer['user_id'], 'crypto_id' => $crypto]
+                    );
+                    $insert = DB::insert("INSERT into cryptoBalance (`balance`, `user_id`, `crypto_id`, `c_t_id`) values (?, ?, ?, ?)", [
+                        $changeofferamount + $offer['balance'],
+                        $offer['user_id'],
+                        $crypto,
+                        $id
+                    ]);
+                    if ($insert) {
+                        DB::update("UPDATE `users` set `balance` = ? where `id` = ?", [
+                            $offer['user_balance'] + ($changeofferamount * ($price - round(($price + $offer['price']) / 2))),
+                            $offer['user_id']
+                        ]);
+                        if ($changeofferamount > $amount) {
+                            DB::update("UPDATE `offers` set `amount` = ? where `id` = ?", [
+                                $offer['amount'] - $changeofferamount,
+                                $offer['id']
+                            ]);
+                        } else {
+                            DB::delete("DELETE from `offers` where `id` = ?", [
+                                $offer['id']
+                            ]);
+                        }
+                        $traded += $changeofferamount * round(($price + $offer['price']) / 2);
+                        $amount -= $changeofferamount;
+                    }
+                }
+            } else {
+                $offer = DB::select("SELECT `offers`.`id`,`offers`.`price`,`offers`.`amount`,`offers`.`user_id`,`users`.`balance` from `offers`
+                left join `users`
+                on `users`.`id` = `offers`.`user_id`
+                where `offers`.`crypto_id` = ? and `offers`.`selling` = ? and `offers`.`user_id` != ? and `offers`.`price` $sort1 ? order by `offers`.`price` $sort2 limit 1", [
+                $crypto,
+                !$selling,
+                $user,
+                $price,
+                ]);
+                $count = count($offer);
+                if ($count > 0) {
+                    $offer = $offer[0];
+                    $changeofferamount = $offer['amount'];
+                    if ($changeofferamount > $amount) {
+                        $changeofferamount = $amount;
+                    }
+                    $insert = DB::update("UPDATE `users` set `balance` = ? where `id` = ?", [
+                        $offer['balance'] + ($changeofferamount * round(($price + $offer['price']) / 2)),
+                        $offer['user_id']
+                    ]);
+                    if ($insert > 0) {
+                        if ($changeofferamount > $amount) {
+                            DB::update("UPDATE `offers` set `amount` = ? where `id` = ?", [
+                                $offer['amount'] - $changeofferamount,
+                                $offer['id']
+                            ]);
+                        } else {
+                            DB::delete("DELETE from `offers` where `id` = ?", [
+                                $offer['id']
+                            ]);
+                        }
+                        $traded += $changeofferamount;
+                        $back += $changeofferamount * ($price - round(($price + $offer['price']) / 2));
+                        $amount -= $changeofferamount;
+                    }
+                }
+            }
+        }
+        if ($selling) {
+            DB::update("UPDATE `users` set `balance` = `balance` + ? where `id` = ?", [
+                $traded,
+                $user
+            ]);
+            if ($amount > 0) {
+                if(isset($offer_id)) {
+                    DB::update("UPDATE `offers` set `price` = ? ,`amount` = ?, `selling` = ? where `id` = ?", [
+                        $price,
+                        $amount,
+                        $selling,
+                        $offer_id,
+                    ]);
+                } else {
+                    DB::insert("INSERT into `offers` (`user_id`, `crypto_id`, `price`, `amount`, `selling`) values (?, ?, ?, ?, ?)", [
+                        $user,
+                        $crypto,
+                        $price,
+                        $amount,
+                        $selling
+                    ]);
+                }
+            } elseif (isset($offer_id)) {
+                DB::delete("DELETE from `offers` where `id` = ?", [
+                    $offer_id,
+                ]);
+            }
+        } else {
+            $balance = DB::select("SELECT cryptoBalance.balance from (
+                SELECT user_id, MAX(`c_t_id`) as `CT`
+                FROM cryptoBalance
+                WHERE crypto_id = ? and user_id = ?
+                GROUP BY user_id
+            ) r
+            left join `cryptoBalance`
+            on `cryptoBalance`.`user_id` = r.`user_id` and `cryptoBalance`.`c_t_id` = r.`CT`
+            where `cryptoBalance`.crypto_id = ?", [
+                $crypto,
+                $user,
+                $crypto,
+            ]);
+            $balance = isset($balance[0]['balance']) ? $balance[0]['balance'] : 0;
+            $id = DB::table('crypto_transactions')->insertGetId(
+                ['in_out' => true, 'amount' => $traded, 'state' => 'success',
+                'user_id' => $user, 'crypto_id' => $crypto]
+            );
+            DB::insert("INSERT into cryptoBalance (`balance`, `user_id`, `crypto_id`, `c_t_id`) values (?, ?, ?, ?)", [
+                $traded + $balance,
+                $user,
+                $crypto,
+                $id
+            ]);
+            DB::update("UPDATE `users` set `balance` = `balance` + ? where `id` = ?", [
+                $back,
+                $user
+            ]);
+            if ($amount > 0) {
+                if(isset($offer_id)) {
+                    DB::update("UPDATE `offers` set `price` = ? ,`amount` = ?, `selling` = ? where `id` = ?", [
+                        $price,
+                        $amount,
+                        $selling,
+                        $offer_id,
+                    ]);
+                } else {
+                    DB::insert("INSERT into `offers` (`user_id`, `crypto_id`, `price`, `amount`, `selling`) values (?, ?, ?, ?, ?)", [
+                        $user,
+                        $crypto,
+                        $price,
+                        $amount,
+                        $selling
+                    ]);
+                }
+            } elseif (isset($offer_id)) {
+                DB::delete("DELETE from `offers` where `id` = ?", [
+                    $offer_id,
+                ]);
+            }
+        }
+        return response()->json([
+            'status' => 200
+        ]); 
     }
 
     /**
@@ -68,6 +249,7 @@ class OfferController extends Controller
      */
     public function store(string $crypto, StoreOfferRequest $request)
     {
+        $user_id = $request->user()->id;
         $cryptos = DB::select("SELECT `slug`,`id` from cryptos where slug = '?'", [
             $crypto,
         ]);
@@ -118,6 +300,7 @@ class OfferController extends Controller
         ]);
         abort_unless(count($offers) > 0, 401, 'Offer not found');
         $off = $offers[0];
+        abort_unless($request->user()->id == $off['user_id'], 403, 'Access denied');
         $validated = $request->validated();
         return $this->TradeOffer($cryptos[0]['id'], $off['user_id'], $validated['price'], $validated['amount'], $off['selling'], $off['id']);
     }
@@ -125,12 +308,22 @@ class OfferController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $crypto, string $offer)
+    public function destroy(Request $request ,string $crypto, string $offer)
     {
+        $cryptos = DB::select("SELECT `slug`,`id` from cryptos where slug = '?'", [
+            $crypto,
+        ]);
+        abort_unless(count($cryptos) > 0, 404, 'Crypto not found');
+        $offers = DB::select("SELECT `id` from `offers` where `crypto_id` = ? and `id` = ?", [
+            $cryptos[0]['id'],
+            $offer
+        ]);
+        abort_unless(count($offers) > 0, 401, 'Offer not found');
+        $off = $offers[0];
+        abort_unless($request->user()->id == $off['user_id'], 403, 'Access denied');
         $insert = DB::delete("DELETE from offers where id = ?", [
             $offer,
         ]);
-        abort_unless($insert > 0, 404, 'Offer not found');
         return response()->json([
             'status' => 200,
             'result' => $insert
